+++ Assets/Scripts/Minigames/PhoneMinigameController.cs
@@ -111,7 +111,7 @@ public class PhoneMinigameController : MonoBehaviour {
         LoggingService.Log(LoggingService.LogCategory.Minigame, "{\"message\":\"Phone minigame completed\",\"score\":" + score + "}");
         Store.Instance.minigameScore = score;
         int scoreForStore = score == 0 ? 0b000 : score == 1 ? 0b100 : 0b110;
-        Store.Instance.SetLevelScore((int)Store.Level.Phone, scoreForStore);
+        Store.Instance.SetLevelScore(Store.Level.Phone, scoreForStore);
         Store.Instance.quizToLoad = Store.Quiz.Phone;
 
         DOVirtual.DelayedCall(2, () => SceneManager.LoadScene("Quiz"));
